requirements.txt
absl-py==2.1.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cvzone==1.6.1
cycler==0.12.1
Django==5.0.4
flatbuffers==24.3.25
fonttools==4.51.0
humanfriendly==10.0
idna==3.7
imageio==2.34.0
jax==0.4.26
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.42.0
matplotlib==3.8.4
mediapipe==0.10.11
ml-dtypes==0.4.0
mpmath==1.3.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
onnxruntime==1.17.3
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
protobuf==3.20.3
pycparser==2.22
pygame==2.5.2
PyMatting==1.1.12
pyparsing==3.1.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
referencing==0.34.0
rembg==2.0.56
requests==2.31.0
rpds-py==0.18.0
scikit-image==0.23.1
scipy==1.13.0
six==1.16.0
sounddevice==0.4.6
sqlparse==0.5.0
sympy==1.12
tifffile==2024.2.12
tqdm==4.66.2
tzdata==2024.1
urllib3==2.2.1